+++ src/components/components.zig
@@ -67,6 +67,9 @@ pub const alert = @import("alert.zig");
 pub const group_box = @import("group_box.zig");
 pub const description_list = @import("description_list.zig");
 pub const searchable_list = @import("searchable_list.zig");
+pub const sidebar = @import("sidebar.zig");
+pub const dock = @import("dock.zig");
+pub const tiles = @import("tiles.zig");
 pub const tree = @import("tree.zig");
 pub const color_picker = @import("color_picker.zig");